oh, there's t_regression.c now, neat.
